Moved SessionLoadedIdlingResource.kt to the new idlingResources Package.
