+++ app/src/androidTest/java/org/mozilla/focus/activity/AccessAboutAndYourRightsPagesTest.kt
@@ -22,7 +22,7 @@ import org.junit.runner.RunWith
 import org.mozilla.focus.R
 import org.mozilla.focus.helpers.EspressoHelper.openMenu
 import org.mozilla.focus.helpers.MainActivityFirstrunTestRule
-import org.mozilla.focus.helpers.SessionLoadedIdlingResource
+import org.mozilla.focus.idlingResources.SessionLoadedIdlingResource
 
 // This test visits each page and checks whether some essential elements are being displayed
 // https://testrail.stage.mozaws.net/index.php?/cases/view/81664
