+++ app/src/androidTest/java/org/mozilla/focus/idlingResources/SessionLoadedIdlingResource.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.focus.helpers
+package org.mozilla.focus.idlingResources
 
 import androidx.test.espresso.IdlingResource
 import androidx.test.platform.app.InstrumentationRegistry
